ext4: Fix offset overflow on 32-bit archs in ext4_iomap_begin() (bsc#1079747).
suse-commit: bbceb807e6c4c100d30e6d62092dca003f4632d8

+++ fs/ext4/inode.c
@@ -3442,7 +3442,7 @@ retry:
 		iomap->flags |= IOMAP_F_DIRTY;
 	iomap->bdev = inode->i_sb->s_bdev;
 	iomap->dax_dev = sbi->s_daxdev;
-	iomap->offset = first_block << blkbits;
+	iomap->offset = (u64)first_block << blkbits;
 
 	if (ret == 0) {
 		iomap->type = IOMAP_HOLE;
